requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
requests==2.31.0
python-multipart==0.0.6
google-genai==0.3.2
google-cloud-aiplatform==1.38.1
google-auth==2.23.4
chromadb==0.4.22
numpy==1.24.3
pandas==2.0.3 